Line & item update + dev tools changes
A comment on Planet Vampire made me aware of a changed line that could
be restored by creating a new item, which I did. Also I decided to
remove the Python update as it's creator mentioned on PV that it doesn't
really work, and so I included the new map sources into the extras area
instead. I also renamed other files there and added two trailer videos
that show content that the patch restored!